Bug 1412808 - Remove the LITERAL helper. (#25)
Its most common use-case can be fullfilled with a simple
FTL.TextElement() because CONCAT is also capable of concatenating
TextElements and Placables.

Its other use-case (defining a hard-coded value for a Message) is rare
and the alternative is not too bad:

    FTL.Pattern([
        FTL.TextElement("Hardcoded value")
    ])

Its third use-case was in REPLACE as the foreach function. That's
already covered by the default value of foreach, however.