Allow CONCAT(*elements) and CONCAT(elements=[…])
CONCAT used to override FTL.BaseNode.traverse and FTL.BaseNode.to_json because
of its use of the *elements varargs in __init__.

Overriding to_json isn't required because the base mathod already supports
lists.

For traverse, we allow CONCAT to be constructed with both the *elements
varargs (useful when writing migration specs) and with the elements=[…] keyword
arg (used by traverse because elements is an attribute of CONCAT instances).