Never migrate partial translations (#44)
Partial translations may break the AST because they produce `TextElements` with
`None` values. For now, explicitly skip any transforms which depend on at least
one missing legacy string.

In the future we might be able to allow partial migrations in some situations,
like migrating a subset of attributes of a message. See
https://bugzilla.mozilla.org/show_bug.cgi?id=1321271.