Fix Comment span to include the terminal line break
Port of https://github.com/projectfluent/fluent.js/pull/257.

Comments' content includes the terminal line break and the span should
reflect it. The end span should be the offset of the beginning of the
next line following the Comment, not the offset of the Comment's
terminal line break.

This also means that we can remvoe the special case in skipBlankLines.
The meaning changes slightly: skipBlankLines now return the number of
all line breaks skipped, regardless of whether they belonged to
fully-blank lines or not. This is OK for our use-case, however.