Make resolver threadsaf(er) again, fix isolation
This patch removes state from the compiled tree.
For one, the recursion detection on patterns keeps its state in
the `ResolverEnvironment` now, instead of in the compiled tree.
Also, the question of bidi isolation is now kept in the resolver
context, instead of in the compiled tree.
For single placeables, which don't get bidi-isolated ever, there's
now a distinct NeverIsolatingPlaceable, which also takes care of
resolving fluent values to strings.

This should fix the test cases in #108 and #109