Support type parameter (ordinal/cardinal) in NUMBER()
The type option is simply ignored outside of a select expression, and
allows to select based on CLDR plural categories for ordinals (which
differ from normal plurals aka cardinal numbers).

This is as per (part of) the docs, and per javascript implementation,
see projectfluent/fluent#259 -- I haven’t found any definition of
NUMBER() in the spec.