Add a type alias for stream location
Prior to this change, the location in the parser stream was represented
by the current index. We would like to update the parser stream to also
return the source position (row/column) as part of the location
information.

This change adds a type alias to represent a location in the parser
stream, and a property to return it. Currently, this is just the index,
but we will update it in future commits.

This allows us to break up the change to make it easier to review, as we
can switch code paths which create span information to use locations
before making any functional changes.