Pass through source position information in spans
Prior to this change, we were tracking source position information in
the stream, but were not storing it in the resulting spans.

This change adds start_position and end_position SourcePosition
attributes to Span nodes. To make this work we end up complicating the
constructor for Span nodes, so that it can either take a location or an
index/position combo (for JSON deserialization).