Track source position information in Streams
Prior to this change, the only positional variables which were tracked
when moving through the stream was the current index in the stream.

We would like to start reporting richer positional information from the
parser, specifically the position (in terms of row/column) in the source
object.

This change allows this by updating the Stream class to additionally
track its current (zero-indexed) row and column while moving through the
stream. We will persist this information into the syntax tree in the
following commits.