Update tests to support Babel 2.15.0
As part of the Babel 2.15.0 release, an upgrade to CLDR 44 was included.
The new CLDR version changes the date format for the en-GB locale to
remove a comma in the full format.

This change updates the tests so that both the old and new formats can
be handled. This is done instead of pinning babel to a lower version,
since we just care that formatting takes place.

[Babel issue]: https://github.com/python-babel/babel/issues/1070
[CLDR 44 changes]: https://www.unicode.org/cldr/charts/44/delta/en.html#Formats%20-%20Standard%20-%20Date%20Formats