Respect the `with_spans` parser option for annotations (#201)
* Add versions of the syntax tests without spans

Prior to this change, all of the syntax tests covered the behaviour of
the parser when with_spans is set to True.

This change updates the test generation to create a version of each test
which tests the parser when with_spans is set to False. To achieve this,
we strip the span information from the expected file (rather than
needing to maintain two files).

* Respect the with_spans setting for Annotations

Prior to this change, Annotations would always have span information,
irrespective of the value of the with_spans argument to the parser.

This change changes the behaviour to only include span information if
with_spans if set to True.