Fix test failures on Windows
The test failures on Windows (which are not present on Linux) are caused by the test code using simple string compare (through dict) when simulating files. On Windows, paths are formed with "\" instead of "/", which causes inconsistency in the string compare. (Specifically, the "\" were inserted by os.path.join in FluentResourceLoader.)